Move broker client into pkg/brokerapi.
Fixes #157.
