+++ pkg/brokerapi/broker_client.go
@@ -11,7 +11,7 @@ See the License for the specific language governing permissions and
 limitations under the License.
 */
 
-package util
+package brokerapi
 
 import (
 	model "github.com/kubernetes-incubator/service-catalog/model/service_broker"
